Crate simple_democracy

This is a facade over Substrate’s democracy pallet offering similar functionality with a few differences. It also assumes that sending extrinsics to Substrate’s democracy pallet are disallowed and this is configured in runtime/lib.rs This pallet offers a “simple majority” governance, by mostly restricting certain features of Substrate’s democracy pallet. Here only Council can vote but anyone can propose by locking a fixed amount of tokens, PublicProposalDeposit. Similar to Substrate’s governance pallet, both Council and general public take turns proposing, and a proposal follows “propose” -> “referendum” -> “enact” sequence. General public’s proposals can be “seconded” by other token holder. Technical committee can fast track proposal made by Council member. A simple majority of Council members’ votes (>50%) are needed to accept an ongoing referendum. Council members can cancel a proposal made by the public. Setting up of Council and Technical committee can be seen in the pallet’s tests.

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Module

Enums

Call
Dispatchable calls.

Error

Event
Events for this module.

Traits

Trait